+++ client/js/irc/commands/away.ts
@@ -4,6 +4,7 @@
  */
 
 import {formatLine} from "../message";
+import {userAwayChanged} from "../presence";
 import {trailingLine} from "../wire";
 import type {Command} from "../types";
 
@@ -11,10 +12,12 @@ const away: Command = {
 	commands: ["away", "back"],
 	input({client, cmd, args}) {
 		if (cmd === "away") {
+			userAwayChanged(client, true);
 			client.send(trailingLine("AWAY", [args.join(" ") || " "]));
 			return;
 		}
 
+		userAwayChanged(client, false);
 		client.send(formatLine({command: "AWAY", params: []}));
 	},
 };